Fixed typing errors, now trying to have a beautiful merge

+++ temporal_processing.py
@@ -20,11 +20,12 @@ def butter_bandpass_filter(data: NumpyArray, lowcut: float, highcut: float, fs: 
     return y
 
 
-def frequential_filter_video(spatially_processed_frames: NumpyArray, lowcut: float, highcut: float, fs: float, order=5) -> NumpyArray:
+def frequential_filter_video(spatially_processed_frames: NumpyArray, lowcut: float, highcut: float, fs: float,
+                             order=5) -> NumpyArray:
     width, height = get_width_and_height_from_frames(spatially_processed_frames)
     total_time = spatially_processed_frames.shape[0]
     pre_processed_video = np.zeros((3, height, width, total_time))
-    post_processed_video = np.zeros((total_time, height, width, 3))
+    post_processed_video = np.zeros((total_time, height, width, 3), dtype=np.uint8)
     # Shape modification to ease frequential filtering
     for t in range(total_time):
         for i in range(height):
@@ -42,5 +43,5 @@ def frequential_filter_video(spatially_processed_frames: NumpyArray, lowcut: flo
         for i in range(height):
             for j in range(width):
                 for t in range(total_time):
-                    post_processed_video[t][i][j][color] = pre_processed_video[color][i][j][t]
+                    post_processed_video[t][i][j][color] = int(pre_processed_video[color][i][j][t])
     return post_processed_video
